Search people tags

+++ assets/js/profile.js
@@ -239,19 +239,22 @@
       var oThis = this;
 
       $('button#video-delete-btn').click(function(event) {
-        var button = this;
+        var button = this,
+          deleteEntry = window.confirm('Do you want to delete this entry ? ');
 
-        event.preventDefault();
+        if (deleteEntry) {
+          event.preventDefault();
 
-        var videoId = +$(this).attr('data-video-id');
+          var videoId = +$(this).attr('data-video-id');
 
-        var updateButtonStatus = function() {
-          $(button).html('Saved');
-          $(button).addClass('disabled');
-          $(button).css('pointer-events', 'none');
-        };
+          var updateButtonStatus = function() {
+            $(button).html('Saved');
+            $(button).addClass('disabled');
+            $(button).css('pointer-events', 'none');
+          };
 
-        oThis.deleteVideo(videoId, updateButtonStatus);
+          oThis.deleteVideo(videoId, updateButtonStatus);
+        }
       });
     },
